+++ driver/web/apis_admin.go
@@ -720,10 +720,16 @@ func (h AdminApisHandler) createApplicationGroup(l *logs.Log, r *http.Request, c
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
 	}
-	var requestData Def.AdminReqCreateApplicationGroup
+	var requestData Def.AdminReqApplicationGroup
 	err = json.Unmarshal(data, &requestData)
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeAppOrgGroup, nil, err, http.StatusBadRequest, true)
+	}
+
+	//system flag
+	system := false
+	if requestData.System != nil {
+		system = *requestData.System
 	}
 
 	//permissions names
@@ -738,10 +744,80 @@ func (h AdminApisHandler) createApplicationGroup(l *logs.Log, r *http.Request, c
 		rolesIDs = *requestData.Roles
 	}
 
+	//account ids
+	var accountIDs []string
+	if requestData.AccountIds != nil {
+		accountIDs = *requestData.AccountIds
+	}
+
+	if len(permissionNames) == 0 && len(rolesIDs) == 0 {
+		return l.HttpResponseErrorData(logutils.StatusMissing, "permissions and application organization roles", nil, nil, http.StatusBadRequest, false)
+	}
+
 	assignerPermissions := strings.Split(claims.Permissions, ",")
-	group, err := h.coreAPIs.Administration.AdmCreateAppOrgGroup(requestData.Name, permissionNames, rolesIDs, claims.AppID, claims.OrgID, assignerPermissions, claims.System, l)
+	group, err := h.coreAPIs.Administration.AdmCreateAppOrgGroup(requestData.Name, requestData.Description, system, permissionNames, rolesIDs, accountIDs, claims.AppID, claims.OrgID, assignerPermissions, claims.System, l)
 	if err != nil || group == nil {
-		return l.HttpResponseErrorAction(logutils.ActionGet, model.TypeAppOrgGroup, nil, err, http.StatusInternalServerError, true)
+		return l.HttpResponseErrorAction(logutils.ActionCreate, model.TypeAppOrgGroup, nil, err, http.StatusInternalServerError, true)
+	}
+
+	respGroup := appOrgGroupToDef(*group)
+	response, err := json.Marshal(respGroup)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionMarshal, model.TypeAppOrgGroup, nil, err, http.StatusInternalServerError, false)
+	}
+	return l.HttpResponseSuccessJSON(response)
+}
+
+// updateApplicationGroup updates an application group
+func (h AdminApisHandler) updateApplicationGroup(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	params := mux.Vars(r)
+	groupID := params["id"]
+	if len(groupID) <= 0 {
+		return l.HttpResponseErrorData(logutils.StatusMissing, logutils.TypeQueryParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
+	}
+
+	data, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
+	}
+	var requestData Def.AdminReqApplicationGroup
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeAppOrgGroup, nil, err, http.StatusBadRequest, true)
+	}
+
+	//system flag
+	system := false
+	if requestData.System != nil {
+		system = *requestData.System
+	}
+
+	//permissions names
+	var permissionNames []string
+	if requestData.Permissions != nil {
+		permissionNames = *requestData.Permissions
+	}
+
+	//roles ids
+	var rolesIDs []string
+	if requestData.Roles != nil {
+		rolesIDs = *requestData.Roles
+	}
+
+	//account ids
+	var accountIDs []string
+	if requestData.AccountIds != nil {
+		accountIDs = *requestData.AccountIds
+	}
+
+	if len(permissionNames) == 0 && len(rolesIDs) == 0 {
+		return l.HttpResponseErrorData(logutils.StatusMissing, "permissions and application organization roles", nil, nil, http.StatusBadRequest, false)
+	}
+
+	assignerPermissions := strings.Split(claims.Permissions, ",")
+	group, err := h.coreAPIs.Administration.AdmUpdateAppOrgGroup(groupID, requestData.Name, requestData.Description, system, permissionNames, rolesIDs, accountIDs, claims.AppID, claims.OrgID, assignerPermissions, claims.System, l)
+	if err != nil || group == nil {
+		return l.HttpResponseErrorAction(logutils.ActionUpdate, model.TypeAppOrgGroup, nil, err, http.StatusInternalServerError, true)
 	}
 
 	respGroup := appOrgGroupToDef(*group)