+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Update group admin API [#518](https://github.com/rokwire/core-building-block/issues/518)
 
 ## [1.26.0] - 2022-10-05
 ### Fixed